feat: move SpecParser to spec.rs
